tools/screenshot_manager/requirements.txt
watchdog
psutil
Pillow
